Fix runtime logging in parallel_tests for turnip
When rutabaga was included, the monkeypatch was confusing
rspec when finding the file of the example group. This was
then not being output in the corresponding runtime log.
