+++ lib/rutabaga.rb
@@ -2,5 +2,6 @@ require 'rutabaga/version'
 require 'turnip'
 require 'rutabaga/feature'
 require 'rutabaga/example_group/feature'
+require 'rspec/core/example_group_patch'
 require 'rutabaga/util'
 require 'rutabaga/turnip' # is disabled by 'rutabaga/no_turnip'
